Moves user views to class view
