+++ bookwyrm/templates/layout.html
@@ -82,7 +82,7 @@
                         </a>
                     </li>
                     <li>
-                        <a href="/user-edit" class="navbar-item">
+                        <a href="/edit-profile" class="navbar-item">
                             Settings
                         </a>
                     </li>
