+++ bookwyrm/templates/user.html
@@ -7,7 +7,7 @@
     </div>
     {% if is_self %}
     <div class="column is-narrow">
-        <a href="/user-edit/">
+        <a href="/edit-profile">
             <span class="icon icon-pencil">
                 <span class="is-sr-only">Edit profile</span>
             </span>
