+++ bookwyrm/tests/test_views.py
@@ -221,83 +221,6 @@ class Views(TestCase):
             response.context_data['user_results'][0], self.local_user)
 
 
-    def test_import_page(self):
-        ''' there are so many views, this just makes sure it LOADS '''
-        request = self.factory.get('')
-        request.user = self.local_user
-        result = views.import_page(request)
-        self.assertIsInstance(result, TemplateResponse)
-        self.assertEqual(result.template_name, 'import.html')
-        self.assertEqual(result.status_code, 200)
-
-
-    def test_import_status(self):
-        ''' there are so many views, this just makes sure it LOADS '''
-        import_job = models.ImportJob.objects.create(user=self.local_user)
-        request = self.factory.get('')
-        request.user = self.local_user
-        with patch('bookwyrm.tasks.app.AsyncResult') as async_result:
-            async_result.return_value = []
-            result = views.import_status(request, import_job.id)
-        self.assertIsInstance(result, TemplateResponse)
-        self.assertEqual(result.template_name, 'import_status.html')
-        self.assertEqual(result.status_code, 200)
-
-
-    def test_user_page(self):
-        ''' there are so many views, this just makes sure it LOADS '''
-        request = self.factory.get('')
-        request.user = self.local_user
-        with patch('bookwyrm.views.is_api_request') as is_api:
-            is_api.return_value = False
-            result = views.user_page(request, 'mouse')
-        self.assertIsInstance(result, TemplateResponse)
-        self.assertEqual(result.template_name, 'user.html')
-        self.assertEqual(result.status_code, 200)
-
-        with patch('bookwyrm.views.is_api_request') as is_api:
-            is_api.return_value = True
-            result = views.user_page(request, 'mouse')
-        self.assertIsInstance(result, ActivitypubResponse)
-        self.assertEqual(result.status_code, 200)
-
-
-    def test_followers_page(self):
-        ''' there are so many views, this just makes sure it LOADS '''
-        request = self.factory.get('')
-        request.user = self.local_user
-        with patch('bookwyrm.views.is_api_request') as is_api:
-            is_api.return_value = False
-            result = views.followers_page(request, 'mouse')
-        self.assertIsInstance(result, TemplateResponse)
-        self.assertEqual(result.template_name, 'followers.html')
-        self.assertEqual(result.status_code, 200)
-
-        with patch('bookwyrm.views.is_api_request') as is_api:
-            is_api.return_value = True
-            result = views.followers_page(request, 'mouse')
-        self.assertIsInstance(result, ActivitypubResponse)
-        self.assertEqual(result.status_code, 200)
-
-
-    def test_following_page(self):
-        ''' there are so many views, this just makes sure it LOADS '''
-        request = self.factory.get('')
-        request.user = self.local_user
-        with patch('bookwyrm.views.is_api_request') as is_api:
-            is_api.return_value = False
-            result = views.following_page(request, 'mouse')
-        self.assertIsInstance(result, TemplateResponse)
-        self.assertEqual(result.template_name, 'following.html')
-        self.assertEqual(result.status_code, 200)
-
-        with patch('bookwyrm.views.is_api_request') as is_api:
-            is_api.return_value = True
-            result = views.following_page(request, 'mouse')
-        self.assertIsInstance(result, ActivitypubResponse)
-        self.assertEqual(result.status_code, 200)
-
-
     def test_status_page(self):
         ''' there are so many views, this just makes sure it LOADS '''
         status = models.Status.objects.create(
