+++ bookwyrm/views/__init__.py
@@ -6,3 +6,4 @@ from .landing import About, Home, Feed, Discover
 from .notifications import Notifications
 from .direct_message import DirectMessage
 from .import_data import Import, ImportStatus
+from .user import User, EditUser, Followers, Following
