+++ bookwyrm/vviews.py
@@ -159,121 +159,6 @@ def search(request):
 
 @csrf_exempt
 @require_GET
-def user_page(request, username):
-    ''' profile page for a user '''
-    try:
-        user = get_user_from_username(username)
-    except models.User.DoesNotExist:
-        return HttpResponseNotFound()
-
-    if is_api_request(request):
-        # we have a json request
-        return ActivitypubResponse(user.to_activity())
-    # otherwise we're at a UI view
-
-    try:
-        page = int(request.GET.get('page', 1))
-    except ValueError:
-        page = 1
-
-    shelf_preview = []
-
-    # only show other shelves that should be visible
-    shelves = user.shelf_set
-    is_self = request.user.id == user.id
-    if not is_self:
-        follower = user.followers.filter(id=request.user.id).exists()
-        if follower:
-            shelves = shelves.filter(privacy__in=['public', 'followers'])
-        else:
-            shelves = shelves.filter(privacy='public')
-
-    for user_shelf in shelves.all():
-        if not user_shelf.books.count():
-            continue
-        shelf_preview.append({
-            'name': user_shelf.name,
-            'local_path': user_shelf.local_path,
-            'books': user_shelf.books.all()[:3],
-            'size': user_shelf.books.count(),
-        })
-        if len(shelf_preview) > 2:
-            break
-
-    # user's posts
-    activities = get_activity_feed(
-        request.user,
-        ['public', 'unlisted', 'followers'],
-        queryset=models.Status.objects.filter(user=user)
-    )
-    paginated = Paginator(activities, PAGE_LENGTH)
-    activity_page = paginated.page(page)
-
-    prev_page = next_page = None
-    if activity_page.has_next():
-        next_page = '/user/%s/?page=%d' % \
-                (username, activity_page.next_page_number())
-    if activity_page.has_previous():
-        prev_page = '/user/%s/?page=%d' % \
-                (username, activity_page.previous_page_number())
-    data = {
-        'title': user.name,
-        'user': user,
-        'is_self': is_self,
-        'shelves': shelf_preview,
-        'shelf_count': shelves.count(),
-        'activities': activity_page.object_list,
-        'next': next_page,
-        'prev': prev_page,
-    }
-
-    return TemplateResponse(request, 'user.html', data)
-
-
-@csrf_exempt
-@require_GET
-def followers_page(request, username):
-    ''' list of followers '''
-    try:
-        user = get_user_from_username(username)
-    except models.User.DoesNotExist:
-        return HttpResponseNotFound()
-
-    if is_api_request(request):
-        return ActivitypubResponse(user.to_followers_activity(**request.GET))
-
-    data = {
-        'title': '%s: followers' % user.name,
-        'user': user,
-        'is_self': request.user.id == user.id,
-        'followers': user.followers.all(),
-    }
-    return TemplateResponse(request, 'followers.html', data)
-
-
-@csrf_exempt
-@require_GET
-def following_page(request, username):
-    ''' list of followers '''
-    try:
-        user = get_user_from_username(username)
-    except models.User.DoesNotExist:
-        return HttpResponseNotFound()
-
-    if is_api_request(request):
-        return ActivitypubResponse(user.to_following_activity(**request.GET))
-
-    data = {
-        'title': '%s: following' % user.name,
-        'user': user,
-        'is_self': request.user.id == user.id,
-        'following': user.following.all(),
-    }
-    return TemplateResponse(request, 'following.html', data)
-
-
-@csrf_exempt
-@require_GET
 def status_page(request, username, status_id):
     ''' display a particular status (and replies, etc) '''
     try:
@@ -326,22 +211,6 @@ def replies_page(request, username, status_id):
         return HttpResponseNotFound()
 
     return ActivitypubResponse(status.to_replies(**request.GET))
-
-
-@login_required
-@require_GET
-def edit_profile_page(request):
-    ''' profile page for a user '''
-    user = request.user
-
-    form = forms.EditUserForm(instance=request.user)
-    data = {
-        'title': 'Edit profile',
-        'form': form,
-        'user': user,
-    }
-    return TemplateResponse(request, 'edit_user.html', data)
-
 
 @require_GET
 def book_page(request, book_id):
